consumerPy/requirements.txt
kafka-python>=1.3.5
grpcio>=1.27.2
protobuf>=3.11.3
python-snappy
